feat(REGISTER_PAGE): add Register's flow and adjusting architecture
