+++ front/src/domain/entities/User.ts
@@ -2,5 +2,8 @@ export type User = {
   name: string;
   cpf: string;
   email: string;
+  phone: string;
+  birthdate: string;
   password: string;
+  confirmPassword: string;
 };
